Rename it to casino crate
